Split rest react-client helpers
