+++ .sampo/changesets/issue-414-react-client-boundaries.md
@@ -1,0 +1,5 @@
+---
+npm/@wp-typia/rest: patch
+---
+
+Split the `@wp-typia/rest` React client cache-key, cache-state, and utility helpers into focused modules while keeping the public `react-client` surface stable.
